move base colours to constants/colours.ts
